+++ pyobjc-core/Modules/objc/pyobjc.h
@@ -150,9 +150,11 @@
  */
 
 /* module.m */
-extern PyObject* PyObjC_TypeStr2CFTypeID;
-extern PyObject* PyObjC_callable_docstr_get(PyObject* callable, void* closure);
-extern PyObject* PyObjC_callable_signature_get(PyObject* callable, void* closure);
-extern PyObject* PyInit__objc(void);
+NS_ASSUME_NONNULL_BEGIN
+extern PyObject* _Nullable PyObjC_TypeStr2CFTypeID;
+extern PyObject* _Nullable PyObjC_callable_docstr_get(PyObject* callable, void* _Nullable closure);
+extern PyObject* _Nullable PyObjC_callable_signature_get(PyObject* callable, void* _Nullable closure);
+extern PyObject* _Nullable PyInit__objc(void);
+NS_ASSUME_NONNULL_END
 
 #endif /* PyObjC_H */