+++ pyobjc-core/Modules/objc/struct-sockaddr.h
@@ -4,7 +4,11 @@
 #ifndef PyObjC_STRUCT_SOCKADDR
 #define PyObjC_STRUCT_SOCKADDR
 
+NS_ASSUME_NONNULL_BEGIN
+
 extern int       PyObjC_SockAddrFromPython(PyObject*, void*);
-extern PyObject* PyObjC_SockAddrToPython(void*);
+extern PyObject* _Nullable PyObjC_SockAddrToPython(void*);
+
+NS_ASSUME_NONNULL_END
 
 #endif /*PyObjC_STRUCT_SOCKADDR */